Add shorten-string transform
This transform returns up to the first X characters of the specified string. This is useful for situations where the string may contain fewer characters than the specified length, which currently causes an error with the existing `substring` primitive.

To use, modify the attribute being referenced to point to the desired target, and update the $length value specified on Line 11 to reflect the desired length.